offline: Remove the offline only client
