+++ tests/CMakeLists.txt
@@ -224,7 +224,7 @@ add_aktualizr_test(NAME aklite_offline
 )
 aktualizr_source_file_checks(aklite_offline_test.cc)
 target_include_directories(t_aklite_offline PRIVATE ${TEST_INCS} ${AKTUALIZR_DIR}/tests/ ${AKTUALIZR_DIR}/src/)
-target_link_libraries(t_aklite_offline ${MAIN_TARGET_LIB} ${TEST_LIBS} testutilities uptane_generator_lib ${AKLITE_OFFLINE_LIB} fstatvfs-mock)
+target_link_libraries(t_aklite_offline ${MAIN_TARGET_LIB} ${TEST_LIBS} testutilities uptane_generator_lib fstatvfs-mock)
 set_tests_properties(test_aklite_offline PROPERTIES ENVIRONMENT "LD_PRELOAD=${PROJECT_BINARY_DIR}/tests/libfstatvfs-mock.so")
 set_tests_properties(test_aklite_offline PROPERTIES LABELS "aklite:offline")
 
@@ -236,7 +236,7 @@ add_aktualizr_test(NAME aklite_offline_api
 )
 aktualizr_source_file_checks(aklite_offline_api_test.cc)
 target_include_directories(t_aklite_offline_api PRIVATE ${AKLITE_DIR}/include ${TEST_INCS} ${AKTUALIZR_DIR}/tests/ ${AKTUALIZR_DIR}/src/)
-target_link_libraries(t_aklite_offline_api ${MAIN_TARGET_LIB} ${TEST_LIBS} testutilities uptane_generator_lib ${AKLITE_OFFLINE_LIB} fstatvfs-mock)
+target_link_libraries(t_aklite_offline_api ${MAIN_TARGET_LIB} ${TEST_LIBS} testutilities uptane_generator_lib fstatvfs-mock)
 set_tests_properties(test_aklite_offline_api PROPERTIES ENVIRONMENT "LD_PRELOAD=${PROJECT_BINARY_DIR}/tests/libfstatvfs-mock.so")
 set_tests_properties(test_aklite_offline_api PROPERTIES LABELS "aklite:offline-api")
